requirements.txt
albumentations==1.0.3
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
catalyst==21.12
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.10
cloudpickle==2.0.0
cryptography==36.0.1
cycler==0.11.0
cytoolz==0.11.2
dask==2022.1.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
efficientnet-pytorch==0.6.3
entrypoints==0.3
fastapi==0.71.0
fonttools==4.25.0
fsspec==2022.1.0
hydra-slayer==0.4.0
idna==3.1
imagecodecs==2021.11.20
imageio==2.13.5
imgaug==0.4.0
importlib-metadata==4.8.2
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.13.3
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
locket==0.2.0
loguru==0.5.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
munch==2.5.0
munkres==1.1.4
nbclassic==0.3.5
nbclient==0.5.3
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.6
numpy==1.22.0
olefile==0.46
onnxruntime==1.7.0
omegaconf==2.1.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.3
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
pretrainedmodels==0.7.4
prometheus-client==0.12.0
prompt-toolkit==3.0.20
protobuf==3.19.3
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.10.0
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.10.0
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
setuptools==58.0.4
Shapely==1.8.0
sip==4.19.13
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tensorboardX==2.2
terminado==0.9.4
testpath==0.5.0
tifffile==2021.11.2
timm==0.4.12
toolz==0.11.2
torch==1.10.1
torchvision==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
wheel==0.37.1
widgetsnbextension==3.5.1
zipp==3.7.0
